autopilot: add server.go
